requirements.txt
requests>=2.28.0
urllib3>=1.26.0
yfinance>=0.2.40
tastytrade>=7.0.0
streamlit>=1.30.0
plotly>=5.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
PyYAML>=6.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
